chore: rename power_2.rs -> power.rs
