+++ src/spartan/polys/mod.rs
@@ -3,5 +3,5 @@ pub(crate) mod eq;
 pub(crate) mod identity;
 pub(crate) mod masked_eq;
 pub(crate) mod multilinear;
-pub(crate) mod power2;
+pub(crate) mod power;
 pub(crate) mod univariate;
