+++ src/spartan/polys/power.rs
